# TeamC.SKS.Package.Services/Controllers/LogisticsPartnerApi.cs
/*
 * Parcel Logistics Service
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.3.0
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using TeamC.SKS.Package.Services.Attributes;
using Microsoft.AspNetCore.Authorization;
using TeamC.SKS.Package.Services.DTOs.Models;
using TeamC.SKS.BusinessLogic.Interfaces;
using BL = TeamC.SKS.BusinessLogic.Entities;
using AutoMapper;
using Microsoft.Extensions.Logging;

namespace TeamC.SKS.Package.Services.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class LogisticsPartnerApi : ControllerBase
    {
        private readonly ISenderLogic _senderLogic;
        private readonly ILogisticsPartnerLogic _recLogic;
        private readonly IMapper _mapper;
        private readonly ILogger _logger;
        /// <summary>
        /// 
        /// </summary>
        /// <param name="recLogic"></param>
        /// <param name="sendLogic"></param>
        /// <param name="mapper"></param>
        /// <param name="logger"></param>
        public LogisticsPartnerApi(ILogisticsPartnerLogic recLogic, ISenderLogic sendLogic, IMapper mapper, ILogger<LogisticsPartnerApi> logger)
        {
            _recLogic = recLogic;
            _senderLogic = sendLogic;
            _mapper = mapper;
            _logger = logger;
            _logger.LogTrace("created LogisticsPartnerApi Controller");
        }
        /// <summary>
        /// Transfer an existing parcel from the service of a logistics partner. 
        /// </summary>
        /// <param name="parcel"></param>
        /// <param name="trackingId">The tracking ID of the parcel. E.g. PYJRB4HZ6 </param>
        /// <response code="200">Successfully transitioned the parcel</response>
        /// <response code="400">The operation failed due to an error.</response>
        [HttpPost]
        [Route("/api/parcel/{trackingId}")]
        [ValidateModelState]
        [SwaggerOperation("TransitionParcel")]
        [SwaggerResponse(statusCode: 200, type: typeof(NewParcelInfo), description: "Successfully transitioned the parcel")]
        [SwaggerResponse(statusCode: 400, type: typeof(Error), description: "The operation failed due to an error.")]
        public virtual IActionResult TransitionParcel([FromBody]Parcel parcel, [FromRoute][Required][RegularExpression("^[A-Z0-9]{9}$")]string trackingId)
        {
            try
            {
                var BLmapped = _mapper.Map<BL.Parcel>(parcel);
                BLmapped.TrackingId = trackingId;
                NewParcelInfo newParcelInfo = _senderLogic.SubmitParcelIntoBL(BLmapped);
                //If the trackingId already exists we generate a new one and save the old one into the db
                _recLogic.TransferParcelPartner(newParcelInfo.TrackingId,trackingId);
                
                return StatusCode(200, newParcelInfo);
            }
            catch (BL.BLException exc)
            {
                return handleReturn(System.Reflection.MethodBase.GetCurrentMethod().Name, exc);
            }
            catch (Exception exc)
            {
                return handleReturn(System.Reflection.MethodBase.GetCurrentMethod().Name, exc);
            }
        }
        private IActionResult handleReturn(string method, Exception exception = null)
        {
            if (exception != null)
            {
                _logger.LogError(exception.ToString());
                //throw new ServicesException($"{exception.GetType()} Exception in {method}", exception);
            }
            _logger.LogWarning(method);
            Error err = new Error() { ErrorMessage = $"{method} failed with Message-> {exception.Message}" };
            return BadRequest(err);
        }
    }
}